+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Set up alembic migrations & added migration guide (`alembic/README.md`)
+- Support injecting return values of calldeps into electrons during workflow execution
 
 ## [0.134.0] - 2022-07-15